Add ShowRawPlanFile
The SDK used a non-JSON version of ShowPlanFile to return the
human-friendly, opaque-to-machines plan output during testing. With the
switch to tfexec, this is no longer possible, as all `terraform show`
commands have the `-json` flag applied, meaning all output will be JSON
output.

This adds a new function, `ShowRawPlanFile`, which returns a string of
the output of `terraform show` _without_ the `-json` flag, showing the
raw contents of the plan file. A new function was chosen to avoid
overloading the existing ShowPlanFile, and because the existing
ShowPlanFile returns a `tfjson.Plan` type, which we can't shoehorn a
non-JSON return type into easily. Rather than complicate the return
type, it felt like a new function was a better solution.