remove support for terraform add command
The terraform add command was introduced in certain Terraform v1.1.0 alpha
releases in order to gather feedback. It has now been removed from Core.

Support in tfexec for terraform add was included in the same experimental vein,
with a doc comment indicating that it may be removed. Removal of Add() is
therefore not a breaking change for tfexec.