allow use of local Terraform executable in tests
If the env var TFEXEC_E2ETEST_TERRAFORM_PATH is set to the path of a valid
local Terraform executable, when e2e tests are run, only tests appropriate to
that executable's version will be run.

This involves both the runTest() and runTestVersions() codepaths, since both
are used in e2e tests.