tfcache: update versions
I believe the Latest_v1 constant was always meant to represent the latest available version with a major version of 1, but its use has been ambiguous. If we need a constant to represent the latest available version with a major version of 1 and minor version of 0, it should be named Latest_v1_0. Latest_v1 has been moved to the bottom of the declaration to make this clearer.