add providers mirror command
Add support for `terraform providers mirror`.

This subcommand, added in Terraform 0.13, downloads the providers required for the current configuration to a local directory, whose path is a required positional argument for the command.

The command has two optional flags: -platform=OS_ARCH and -lock-file=true, the latter of which was added in Terraform 1.10.

Co-authored-by: Radek Simko <radeksimko@users.noreply.github.com>
Co-authored-by: Daniel Schmidt <danielmschmidt92@gmail.com>
Co-authored-by: Matej Risek <matej.risek@hashicorp.com>
Co-authored-by: Samsondeen Dare <samsondeen.dare@hashicorp.com>